[FIX] Support RFC-3848 service types in Received headers
